[geometry] Implementation of segments intersection.
