+++ geometry/segment2d.hpp
@@ -6,6 +6,22 @@
 
 namespace m2
 {
+struct IntersectionResult
+{
+  enum class Type
+  {
+    Zero,
+    One,
+    Infinity
+  };
+
+  explicit IntersectionResult(Type type) : m_type(type) { ASSERT_NOT_EQUAL(m_type, Type::One, ()); }
+  explicit IntersectionResult(PointD const & point) : m_point(point), m_type(Type::One) {}
+
+  PointD m_point;
+  Type m_type;
+};
+
 struct Segment2D
 {
   Segment2D() = default;
@@ -17,8 +33,14 @@ struct Segment2D
   PointD m_v;
 };
 
-std::string DebugPrint(Segment2D const & s);
-
 bool IsPointOnSegmentEps(PointD const & pt, PointD const & p1, PointD const & p2, double eps);
 bool IsPointOnSegment(PointD const & pt, PointD const & p1, PointD const & p2);
+
+/// \breif Intersects to segments and finds an intersection point if any.
+/// \note |eps| applies for collinearity and for PointD errors.
+IntersectionResult Intersect(Segment2D const & seg1, Segment2D const & seg2, double eps);
+
+std::string DebugPrint(Segment2D const & s);
+std::string DebugPrint(IntersectionResult::Type type);
+std::string DebugPrint(IntersectionResult const & result);
 }  // namespace m2
